requirements.txt
# AD-Setup Enterprise Python Dependencies

# Core dependencies
PyYAML>=6.0
requests>=2.31.0
docker>=6.1.0
click>=8.1.0

# DNS and domain management
# Note: namecheap doesn't have an official PyPI package
# We'll use requests to interact with their API directly

# Monitoring and metrics
prometheus-client>=0.19.0

# Configuration and environment
python-dotenv>=1.0.0

# Logging and formatting
colorama>=0.4.6
tabulate>=0.9.0

# Security
cryptography>=41.0.0

# Utilities
python-dateutil>=2.8.2
validators>=0.22.0
